fix gallery and magang

+++ app/Http/Controllers/GalleryController.php
@@ -54,18 +54,27 @@ class GalleryController extends Controller
     public function updateImage(Request $request, Gallery $gallery){
 
         $request->validate([
-            "image" => "required|mimes:png,jpg,jpeg|max:4096",
-            "title" => "required"
+            "image" => "mimes:png,jpg,jpeg|max:4096",
+            "title" => "nullable"
         ]);
 
-        Storage::disk('public')->delete(substr($gallery->image, 8));
+        // Hapus gambar lama jika gambar baru diunggah
+        if ($request->hasFile('image')) {
+            Storage::disk('public')->delete(substr($gallery->image, 8));
 
-        $image = $request->file('image');
-        $imageName = time() . "_" . "gallery" . "_" . $image->hashName();
-        $image->storeAs("public/gallery", $imageName);
+            $image = $request->file('image');
+            $imageName = time() . "_" . "gallery" . "_" . $image->hashName();
+            $image->storeAs("public/gallery", $imageName);
 
-        $gallery->title = $request->title;
-        $gallery->image = "storage/gallery/" . $imageName;
+            // Update field image
+            $gallery->image = "storage/gallery/" . $imageName;
+        }
+
+        // Update field title jika ada input baru
+        if ($request->filled('title')) {
+            $gallery->title = $request->title;
+        }
+
         $gallery->save();
 
         return response()->json([
